+++ R/AminoAcids.R
@@ -228,7 +228,7 @@ aliphatic <- function(seq, normalize=TRUE) {
 #' \enumerate{
 #'   \item  Moore DS. Amino acid and peptide net charges: A simple calculational procedure. 
 #'            Biochem Educ. 13, 10-11 (1985).
-#'   \item  \url{http://emboss.sourceforge.net/apps/cvs/emboss/apps/iep.html}
+#'   \item  \url{https://emboss.sourceforge.net/apps/cvs/emboss/apps/iep.html}
 #'   }
 #'   
 #' @seealso 
@@ -462,7 +462,7 @@ countPatterns <- function(seq, patterns, nt=TRUE, trim=FALSE, label="region") {
 #'   \item  Wu YC, et al. The relationship between CD27 negative and positive B cell 
 #'            populations in human peripheral blood. 
 #'            Front Immunol 2, 1-12 (2011).
-#'   \item  \url{http://emboss.sourceforge.net/apps/cvs/emboss/apps/iep.html}
+#'   \item  \url{https://emboss.sourceforge.net/apps/cvs/emboss/apps/iep.html}
 #' }
 #' 
 #' @seealso