+++ inst/doc/AminoAcids-Vignette.R
@@ -1,4 +1,4 @@
-## ---- eval=TRUE, warning=FALSE, message=FALSE---------------------------------
+## ----eval=TRUE, warning=FALSE, message=FALSE----------------------------------
 # Load required packages
 library(alakazam)
 library(dplyr)
@@ -7,7 +7,7 @@ library(dplyr)
 data(ExampleDb)
 db <- ExampleDb[ExampleDb$sample_id == "+7d", ]
 
-## ---- eval=TRUE, warning=FALSE, fig.width=7.5, fig.height=6-------------------
+## ----eval=TRUE, warning=FALSE, fig.width=7.5, fig.height=6--------------------
 db_props <- aminoAcidProperties(db, seq="junction", trim=TRUE, 
                                 label="cdr3")
 
@@ -44,12 +44,12 @@ g4 <- ggplot(db_props, aes(x=c_call, y=cdr3_aa_acidic)) + tmp_theme +
 # Plot in a 2x2 grid
 gridPlot(g1, g2, g3, g4, ncol=2)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 db_props <- aminoAcidProperties(db, seq="junction", property=c("gravy", "charge"),
                                 trim=TRUE, label="cdr3")
 dplyr::select(db_props[1:3, ], starts_with("cdr3"))
 
-## ---- eval=TRUE, warning=FALSE, message=FALSE---------------------------------
+## ----eval=TRUE, warning=FALSE, message=FALSE----------------------------------
 # Load the relevant data objects from the seqinr package
 library(seqinr)
 data(aaindex)
@@ -63,7 +63,7 @@ db_props <- aminoAcidProperties(db, seq="junction", property=c("gravy", "charge"
                                 hydropathy=h, pK=p)
 dplyr::select(db_props[1:3, ], starts_with("cdr3"))
 
-## ---- eval=TRUE, warning=FALSE, message=FALSE---------------------------------
+## ----eval=TRUE, warning=FALSE, message=FALSE----------------------------------
 # Translate junction DNA sequences to amino acids and trim first and last codons
 cdr3 <- translateDNA(db$junction[1:3], trim=TRUE)