+++ inst/doc/Files-Vignette.R
@@ -1,4 +1,4 @@
-## ---- eval=TRUE, warning=FALSE, message=FALSE---------------------------------
+## ----eval=TRUE, warning=FALSE, message=FALSE----------------------------------
 # Set the file paths from inside the package directory
 # These files are smaller versions of the example databases previously mentioned
 changeo_file <- system.file("extdata", "example_changeo.tab.gz", package="alakazam")
@@ -8,7 +8,7 @@ airr_file <- system.file("extdata", "example_airr.tsv.gz", package="alakazam")
 db_changeo <- alakazam::readChangeoDb(changeo_file)
 db_airr <- airr::read_rearrangement(airr_file)
 
-## ---- eval=FALSE, warning=FALSE, message=FALSE--------------------------------
+## ----eval=FALSE, warning=FALSE, message=FALSE---------------------------------
 #  # Write the data to a tab-delimited file
 #  alakazam::writeChangeoDb(db_changeo, "changeo.tsv")
 #  airr::write_rearrangement(db_airr, "airr.tsv")