+++ inst/doc/GeneUsage-Vignette.R
@@ -1,4 +1,4 @@
-## ---- eval=TRUE, warning=FALSE, message=FALSE---------------------------------
+## ----eval=TRUE, warning=FALSE, message=FALSE----------------------------------
 # Load required packages
 library(alakazam)
 library(dplyr)
@@ -7,12 +7,12 @@ library(scales)
 # Subset example data
 data(ExampleDb)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Quantify usage at the gene level
 gene <- countGenes(ExampleDb, gene="v_call", groups="sample_id", mode="gene")
 head(gene, n=4)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Assign sorted levels and subset to IGHV1
 ighv1 <- gene %>%
     mutate(gene=factor(gene, levels=sortGenes(unique(gene), method="name"))) %>%
@@ -30,7 +30,7 @@ g1 <- ggplot(ighv1, aes(x=gene, y=seq_freq)) +
     geom_point(aes(color=sample_id), size=5, alpha=0.8)
 plot(g1)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Quantify V family usage by sample
 family <- countGenes(ExampleDb, gene="v_call", groups="sample_id", mode="family")
 
@@ -46,13 +46,13 @@ g2 <- ggplot(family, aes(x=gene, y=seq_freq)) +
     geom_point(aes(color=sample_id), size=5, alpha=0.8)
 plot(g2)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Quantify V family clonal usage by sample and isotype
 family <- countGenes(ExampleDb, gene="v_call", groups=c("sample_id", "c_call"), 
                      clone="clone_id", mode="family")
 head(family, n=4)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Subset to IGHM and IGHG for plotting
 family <- filter(family, c_call %in% c("IGHM", "IGHG"))
 # Plot V family clonal usage by sample and isotype
@@ -68,13 +68,13 @@ g3 <- ggplot(family, aes(x=gene, y=clone_freq)) +
     facet_grid(. ~ c_call)
 plot(g3)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Calculate V family copy numbers by sample and isotype
 family <- countGenes(ExampleDb, gene="v_call", groups=c("sample_id", "c_call"), 
                      mode="family", copy="duplicate_count")
 head(family, n=4)
 
-## ---- eval=TRUE, warning=FALSE------------------------------------------------
+## ----eval=TRUE, warning=FALSE-------------------------------------------------
 # Subset to IGHM and IGHG for plotting
 family <- filter(family, c_call %in% c("IGHM", "IGHG"))
 # Plot V family copy abundance by sample and isotype